One top-level type per file.
