+++ toofz.NecroDancer.Web.Api/Infrastructure/RunsBinder.cs
@@ -1,0 +1,20 @@
+﻿using toofz.NecroDancer.Leaderboards;
+using toofz.NecroDancer.Web.Api.Models;
+
+namespace toofz.NecroDancer.Web.Api.Infrastructure
+{
+    public sealed class RunsBinder : LeaderboardCategoryBaseBinder
+    {
+        public RunsBinder(Categories leaderboardCategories)
+        {
+            category = leaderboardCategories["runs"];
+        }
+
+        readonly Category category;
+
+        protected override LeaderboardCategoryBase GetModel()
+        {
+            return new Runs(category);
+        }
+    }
+}
