+++ toofz.NecroDancer.Web.Api/Models/EnemiesPagination.cs
@@ -1,0 +1,19 @@
+﻿using System.ComponentModel;
+using System.Web.Http.ModelBinding;
+using toofz.NecroDancer.Web.Api.Infrastructure;
+
+namespace toofz.NecroDancer.Web.Api.Models
+{
+    [ModelBinder(BinderType = typeof(PaginationBinder<EnemiesPagination>))]
+    public sealed class EnemiesPagination : IPagination
+    {
+        [MinValue(0)]
+        [MaxValue(int.MaxValue)]
+        [DefaultValue(0)]
+        public int offset { get; set; } = 0;
+        [MinValue(1)]
+        [MaxValue(100)]
+        [DefaultValue(10)]
+        public int limit { get; set; } = 10;
+    }
+}
