+++ toofz.NecroDancer.Web.Api/Models/Replays.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace toofz.NecroDancer.Web.Api.Models
+{
+    public sealed class Replays
+    {
+        public int total { get; set; }
+        public IEnumerable<Replay> replays { get; set; }
+    }
+}
