+++ toofz.NecroDancer.Web.Api/toofz.NecroDancer.Web.Api.csproj
@@ -284,13 +284,42 @@
     <Compile Include="Identity\ApplicationDbContext.Views.cs">
       <DependentUpon>ApplicationDbContext.cs</DependentUpon>
     </Compile>
+    <Compile Include="Infrastructure\CharactersBinder.cs" />
     <Compile Include="Infrastructure\LeaderboardCategoryBaseBinder.cs" />
-    <Compile Include="Infrastructure\LeaderboardCategoryBinders.cs" />
+    <Compile Include="Infrastructure\ModesBinder.cs" />
     <Compile Include="Infrastructure\PaginationBinder.cs" />
+    <Compile Include="Infrastructure\ProductsBinder.cs" />
+    <Compile Include="Infrastructure\RunsBinder.cs" />
     <Compile Include="IQueryableExtensions.cs" />
+    <Compile Include="Models\BulkStore.cs" />
+    <Compile Include="Models\Characters.cs" />
+    <Compile Include="Models\DailyLeaderboard.cs" />
+    <Compile Include="Models\DailyLeaderboardEntries.cs" />
+    <Compile Include="Models\DailyLeaderboards.cs" />
+    <Compile Include="Models\End.cs" />
+    <Compile Include="Models\Enemies.cs" />
+    <Compile Include="Models\EnemiesPagination.cs" />
+    <Compile Include="Models\Enemy.cs" />
+    <Compile Include="Models\Entry.cs" />
     <Compile Include="Models\IPagination.cs" />
-    <Compile Include="Models\LeaderboardCategories.cs" />
+    <Compile Include="Models\Item.cs" />
+    <Compile Include="Models\Items.cs" />
+    <Compile Include="Models\ItemsPagination.cs" />
+    <Compile Include="Models\Leaderboard.cs" />
     <Compile Include="Models\LeaderboardCategoryBase.cs" />
+    <Compile Include="Models\LeaderboardEntries.cs" />
+    <Compile Include="Models\Leaderboards.cs" />
+    <Compile Include="Models\LeaderboardsPagination.cs" />
+    <Compile Include="Models\Modes.cs" />
+    <Compile Include="Models\Player.cs" />
+    <Compile Include="Models\PlayerEntries.cs" />
+    <Compile Include="Models\Players.cs" />
+    <Compile Include="Models\PlayersPagination.cs" />
+    <Compile Include="Models\Products.cs" />
+    <Compile Include="Models\Replay.cs" />
+    <Compile Include="Models\Replays.cs" />
+    <Compile Include="Models\ReplaysPagination.cs" />
+    <Compile Include="Models\Runs.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <WCFMetadata Include="Service References\" />
     <Content Include="Service References\Application Insights\ConnectedService.json" />
@@ -367,8 +396,6 @@
     <Compile Include="Identity\ApplicationDbContext.cs" />
     <Compile Include="Identity\ApplicationUser.cs" />
     <Compile Include="Identity\ApplicationUserManager.cs" />
-    <Compile Include="Models\Objects.cs" />
-    <Compile Include="Models\Parameters.cs" />
     <Compile Include="Models\PlayerModel.cs" />
     <Compile Include="Models\ReplayModel.cs" />
     <Compile Include="Providers\ApplicationOAuthProvider.cs" />
